Fix ODR violation: rename BvhLightBounds in TheRestOfYourLife, fix float_bits.h missing algorithm include, fix Pi->pi in triangle_tests, add CopyPTX Rebuild target. 2149/2149 tests pass.

+++ src/shared/rgb_to_spectrum_table.h
@@ -82,7 +82,7 @@ public:
 		// Integer indices into the table
 		int xi = std::min((int)x, kRes - 2);
 		int yi = std::min((int)y, kRes - 2);
-		int zi = static_cast<int>(FindInterval(kRes, [&](int i) { return zNodes_[i] < z; }));
+		int zi = FindInterval(kRes, [&](int i) { return zNodes_[i] < z; });
 		float dx = x - xi;
 		float dy = y - yi;
 		float dz = (z - zNodes_[zi]) / (zNodes_[zi + 1] - zNodes_[zi]);
